Modified OMOP_match to use SQLAlchemy
The reason I've done this is to have more maintainable queries to the database.
The immediate issue is to make it possible to use 0+ vocabularies. This is hard with string manipulation
Later on, we'll need to allow different domains etc. to be queried. Using an ORM will make this easier

I still need to check how the generated queries work when synonyms are allowed, I'm not sure if it's working how I would like