Updated UI
I've moved the functions called in the UI to a separate file.
Then I've made the main UI entry point a file that allows navigation between two pages.
The first is the text entry as before.
The second page allows the user to upload a .csv file and use one of the columns for input.