Added code for vector search on concepts
The main change I've made is adding embeddings.py in components. This provides a class, Embeddings, which can be called from the app to search a vector database of concepts. If the database doesn't exist, it'll build it. This takes a long time, so the database is saved for subsequent runs.

There's also the possibility of running several different embedding databases if necessary.

I need to test whether the search function actually works, as it's currently building the database, taking about an hour.

I've added an API endpoint for the vector search. It uses the same PipelineRequest object to set the options, so that if we mix it with the LLM pipeline we don't have to change anything.

Added the requirements to the environment.yml, I'll update requirements next.

OMOP_match wasn't using the Concept and ConceptSynonym classes directly, so taken those out

Put anything ending with ".qdrant" in the .gitignore, that way as long as we name our databases with this extension we won't have a problem pushing it to GitHub, which would have an issue with 100s MB being uploaded. We might need to find a way of distributing the embedded store