First attempt at integrating llama-3.1-8b
I've added the latest llama (3.1) to the supported llms.
I also changed how the llama models are loaded, getting rid of the if/else block and replacing it with a dictionary. Partly this is for style, as I think it's just a nicer way of doing it. The other reason is that other arguments to hf_hub_download can optionally be passed when the dictionary value is unpacked.

The problem I get is that the quantized models don't work with the version of llama-cpp-python we have in the requirements.
I upgraded llama-cpp-python to 0.2.87, then the models load, but the output is bad because the chat format passed by haystack is broken at v0.4.0
I'll try upgrading to llama-cpp-haystack 0.4.1