Fixed error with embeddings
This is actually unrelated to this PR, but has to do with how I had implemented the incorporation of vector search. I hadn't tested the UI with the vector search. The problem is that the FastAPI and Streamlit interaction means that streamlit makes an API request with the full JSON with defaults (I think). The problem then is that the enum used to define embedding models is not JSON serializable. I've modified it so the API request now uses a simpler enum that is JSON serializable, but now it uses a dict of EmbeddingModel objects